Big changes to add Eiger "monitor" TIFF support and librarify image server:
1. Add isTiff: utilities for parsing TIFF images produced by the
   Eiger's "monitor" interface.

2. Fix a handful of race conditions and HDF5 handle leaks.

3. Greatly shorten the time that the metadata mutex needs to be
   locked while also fixing a race condition by locking only when
   the pointer to the data is made visible outside of isGetMeta.

4. Remove some of Keith's unfinished redis caching code. We shall
   assume it's broken, and it's not worth figuring out and
   not worth finishing.

Overall the architecture of this Image Server is better off
replaced than extended, but we need the functionality for
rendering images.

TODO: Remove the rest of the redis cache crap, and decouple the
image rendering logic in isReduceImage and isJpeg from the
threading logic so we can use it as a library.